Add files via upload
Hello everyone, in this project I have used the data from Yahoo finance to compare two stocks, RBC and BMO. From 1/21/2022 to 12/21/2022. 
RBC had a 52-week highest of 122.77 USD and a 52-week lowest at 81.57 USD, as compared to BMO, which had a 52-week highest of 116.84 USD and a 52-week lowest of 84.63 USD.
As we compared the volume between those, on 12/13/2022, BMO reached 7273400, whereas RBC volume was at 1235000.